+++ src/lib/util/contextProxy.ts
@@ -6,7 +6,12 @@ export interface ContextProxy extends Omit<CanvasRenderingContext2D, 'canvas'> {
 }
 
 const EXCLUDED_GETTERS = ['drawImage', 'setTransform'];
-const EXCLUDED_SETTERS = ['filter', 'shadowBlur', 'globalCompositeOperation'];
+const EXCLUDED_SETTERS = [
+  'filter',
+  'shadowBlur',
+  'globalCompositeOperation',
+  'globalAlpha'
+];
 const COLOR_OVERRIDES = [
   'drawImage',
   'fill',
@@ -55,14 +60,16 @@ const createContextProxy = (context: CanvasRenderingContext2D) => {
           resizeCanvas();
         }
 
-        if (property === 'drawImage') {
-          proxyContext.fillRect(...(<Parameters<CanvasRect['fillRect']>>args));
-        }
-
         if (COLOR_OVERRIDES.includes(property)) {
           const layerColor = idToRgb(renderingLayerId());
           proxyContext.fillStyle = layerColor;
           proxyContext.strokeStyle = layerColor;
+        }
+
+        if (property === 'drawImage') {
+          proxyContext.fillRect(
+            ...(<Parameters<CanvasRect['fillRect']>>args.slice(1))
+          );
         }
 
         if (!EXCLUDED_GETTERS.includes(property)) {